Backend_Development/Backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
psycopg2-binary==2.8.6